do not attempt to clean up tmp/devices if has never existed

+++ spec/models/device_spec.rb
@@ -43,7 +43,7 @@ RSpec.describe Device, type: :model do
 
   def mk_empty_dir
     newdir = Rails.root.join("tmp").join("devices")
-    FileUtils.remove_entry_secure(newdir)
+    FileUtils.remove_entry_secure(newdir) if Dir.exists?(newdir)
     FileUtils.mkdir_p(newdir)
     newdir
   end
